+++ SistemaEntities/CamionetaEntity.cs
@@ -4,7 +4,7 @@
     {
         public int IdCamioneta { get; set; }
         public string Patente { get; set; }
-        public int TamañoCarga { get; set; }
+        public double TamañoCarga { get; set; }
         public int DistanciaMax { get; set; }
     }
 }
